refactor: migrate all IDs int32 -> int64

+++ src/Crowdin.Api/Distributions/AddDistributionRequest.cs
@@ -20,12 +20,12 @@ namespace Crowdin.Api.Distributions
         
         [JsonProperty("fileIds")]
 #pragma warning disable CS8618
-        public ICollection<int>? FileIds { get; set; }
+        public ICollection<long>? FileIds { get; set; }
 #pragma warning restore CS8618
         
         [JsonProperty("bundleIds")]
 #pragma warning disable CS8618
-        public ICollection<int>? BundleIds { get; set; }
+        public ICollection<long>? BundleIds { get; set; }
 #pragma warning restore CS8618
     }
 }
