+++ src/Crowdin.Api/SourceFiles/FileResource.cs
@@ -11,10 +11,10 @@ namespace Crowdin.Api.SourceFiles
     public class FileInfoResource
     {
         [JsonProperty("id")]
-        public int Id { get; set; }
+        public long Id { get; set; }
         
         [JsonProperty("projectId")]
-        public int ProjectId { get; set; }
+        public long ProjectId { get; set; }
         
         [JsonProperty("branchId")]
         public int? BranchId { get; set; }
@@ -48,7 +48,7 @@ namespace Crowdin.Api.SourceFiles
     public class FileResource : FileInfoResource
     {
         [JsonProperty("revisionId")]
-        public int RevisionId { get; set; }
+        public long RevisionId { get; set; }
         
         [JsonProperty("priority")]
         public Priority Priority { get; set; }
